Fix setting of the timezone in alias naming mode
When in alias naming mode, for most cases, run_command() in the
utilities_common package forwards the call to run_command_in_alias_mode().
This function by default will just spew the output onto the console
instead of returning it. As a result, the caller of run_command() will
get None as a result and the whole command will fail.

The fix is to make run_command_in_alias_mode() return values the same way
run_command() does -- in this case, return the command output and its
exit code as a tuple.

To keep it compatible with the existing behavior, a new return_cmd flag
is accepted as a parameter in run_command_in_alias_mode() and defaults
to False. To enable the new behavior, run_command()'s return_cmd value
is now being forwarded to run_command_in_alias_mode().